fix(chat): split api/chat helpers into utils, soft-hide WIP routes, tighten long-form padding
* fix(api): extract every helper, constant, and type from
  src/routes/api/chat.ts into the new src/utils/chat-api.ts. SolidStart
  was tree-shaking the route file with `?pick=POST` and stripping all
  the `export const guardChatRequest = ...` helpers, producing a runtime
  `ReferenceError: guardChatRequest is not defined` on every POST. The
  route file is now just the POST handler + thin re-exports kept around
  so the existing test imports keep working.
* feat(home-link): new fixed top-right `home.` pill rendered on every
  route except `/`. Single click back to the portfolio from blog/chat/
  brag/nested pages.
* feat(footer): chat. promoted into the FOOTER_LINKS now that the API is
  working end-to-end; brag. intentionally still hidden until the brag
  UX restructure (sub-nav simplification + landing reorder) lands.
* feat(theme-button): only render on `/`. The fixed bottom-right toggle
  was visual noise on long-form routes (blog/chat/brag) where it has
  nothing to balance against.
* feat(noindex): `<Meta name="robots" content="noindex, nofollow" />` on
  /chat and all /brag routes while they iterate. Easy to drop per route
  when ready to invite indexing.
* refactor(brag/nav): normalize trailing slashes when computing the
  active sub-nav pill, so `brag.` lights up immediately on the
  /brag → /brag/ Solid index redirect (was waiting for a click).
* fix(footer): swap `sm:pl-20` → `sm:ml-20` on the `dvf.` stamp so the
  rail-clearance lives outside the link's clickable area instead of
  enlarging the hit target.
* fix(contact-section): pr-12 sm:pr-20 so the heading + GitHub/LinkedIn
  icons sit at the true viewport center on both mobile and desktop
  (symmetric against the nav rail's pl).
* fix(chat-layout): h-[100dvh] instead of min-h-screen on the chat
  flex column so the message list scrolls inside the card instead of
  growing it and pushing the form down the page. Aligned the chat
  wrapper to the long-form route pattern (max-w-7xl, px-4).
* fix(blog-index): inline BlogHeader, drop the awkward fixed-top
  strip + pt-40 spacer hack. BlogHeader's pt-8 → pt-6.
* style(padding): tighten the vertical rhythm on every long-form route:
  py-16 → pt-8 pb-4 on the inner wrapper, pb-28 → pb-8 on the outer,
  footer mt-16 py-8 → mt-8 py-6. ~160px less dead space between content
  and viewport bottom across blog/[slug], brag/*, chat.
* test(theme-button): mock useLocation so the new home-only Show gate
  doesn't crash the suite without a Router.
* test(brag-route): drop assertions for the removed latest_year and
  browse_blog_posts. CTAs.
