+++ src/components/brag/nav.tsx
@@ -24,6 +24,9 @@ const buildBragNavLinks = (): BragNavLink[] => {
   ]
 }
 
+const normalizePath = (path: string) =>
+  path.length > 1 ? path.replace(/\/+$/, '') : path
+
 const BragNav = () => {
   const location = useLocation()
   const links = buildBragNavLinks()
@@ -34,20 +37,23 @@ const BragNav = () => {
       class="flex flex-wrap gap-2 pb-8 border-b border-gray-200 dark:border-gray-800 mb-12"
     >
       <For each={links}>
-        {(link) => (
-          <A
-            href={link.href}
-            class="font-mono text-sm rounded-full border-2 px-4 py-2 transition-all duration-500 hover:scale-105 focus-ring"
-            classList={{
-              'bg-gray-900 text-white border-gray-900 dark:bg-white dark:text-black dark:border-white':
-                location.pathname === link.href,
-              'border-gray-300 dark:border-gray-700 hover:border-blue-500 dark:hover:border-blue-400':
-                location.pathname !== link.href,
-            }}
-          >
-            {link.label}
-          </A>
-        )}
+        {(link) => {
+          const isActive = () => normalizePath(location.pathname) === link.href
+          return (
+            <A
+              href={link.href}
+              class="font-mono text-sm rounded-full border-2 px-4 py-2 transition-all duration-500 hover:scale-105 focus-ring"
+              classList={{
+                'bg-gray-900 text-white border-gray-900 dark:bg-white dark:text-black dark:border-white':
+                  isActive(),
+                'border-gray-300 dark:border-gray-700 hover:border-blue-500 dark:hover:border-blue-400':
+                  !isActive(),
+              }}
+            >
+              {link.label}
+            </A>
+          )
+        }}
       </For>
     </nav>
   )
